### 2017-09-24 #416 @whedon
[PRE REVIEW]: scanstatistics: The R Package
**Submitting author:** @benjak (<a href="http://orcid.org/0000-0002-0927-7183">Benjamin Allévius</a>)
**Repository:** <a href="https://github.com/BenjaK/scanstatistics" target ="_blank">https://github.com/BenjaK/scanstatistics</a>
**Version:** v1.0.0
