### 2017-10-02 #416 @leeper
@robohyun66 Would you be interested in reviewing this for JOSS?
