### 2017-10-30 #416 @leeper
@jsta Would you be interested in reviewing this?
