### 2017-11-30 #416 @leeper
I have thus far been unable to find a reviewer for this. @openjournals/joss-editors - anyone have suggestions or know of willing volunteers?
