### 2018-03-12 #515 @BenjaK
Dear @leeper, I cannot see any activity from @pragyansmita, and while waiting for this review, I have released a new (minor) version of the scanstatistics package. Should I withdraw my submission and submitt anew, so that another reviewer can be assigned, or is anyone else willing to review this submission?
