### 2018-03-15 #515 @pragyansmita
The documentation provided is complete. It includes a good step-by-step example related to New Mexico and occurance of cancer. I will watch out for datasets and problems where this tool can be applied.Very interesting solution and made an engaging read. Thank you.
