### 2018-04-24 #515 @tgerke
Works as advertised, readme was easy to follow, overall excellent package! Only item in the review checklist I didn't fill was the one regarding the LICENSE file, which I was not able to locate. One minor comment related to the statement of need: the spatial problems the package solves are nicely described, but it could be useful to mention key areas this package extends/complements existing software packages that implement a subset of the same methods (e.g. SaTScan).

@leeper Do I need to complete extra steps beyond this comment? This is my first review for JOSS, so learning the process (which, by the way, is fantastic thus far; you all really make this easy).
