### 2018-04-24 #515 @leeper
@tgerke Thank you! This looks great. Really appreciate the review. Since it sounds like you're mostly happy, that's all we'll need. Excellent job on your first review!

@BenjaK Review is in and all looks good. We're nearly there. Two final things:

 - Consider updating [the paper](https://github.com/BenjaK/scanstatistics/blob/master/JOSS/paper.md) regarding statement of need. @tgerke's suggestion seems useful.
 - Please add DOIs for all cited articles (if they have them) in [your .bib file](https://github.com/BenjaK/scanstatistics/blob/master/JOSS/paper.bib)
