### 2018-05-02 #515 @BenjaK
Thank you @tgerke and @leeper. I have now

1. added a file `LICENCE.md` to clarify which license the package goes under (the licence is also found in the `DESCRIPTION` file),
2. updated the statement of need with a reference to SaTScan and exemplified what sets my package apart, and
3. added DOIs for cited articles. Note: One reference has been updated since it is now in press.
