### 2018-05-02 #515 @leeper
@BenjaK Great. Thank you. Two final things:

1. I might further suggest that in the paper you situate the references in the actual paper so that it is clear why you are citing the various papers. While JOSS articles are meant to be short, they are still articles that follow a narrative structure.
2. Can you please issue a DOI for an archived version of the software using [zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or a similar service? Please post the DOI for the archive here once you have it.
