### 2018-05-03 #515 @BenjaK
@leeper I'm in the process of fixing your point 2. I wonder, would it be possible to update the version number of my package submitted to JOSS? It is currently v1.0.0, but I would like it to be v1.0.1. The latter is the version showed per default on the GitHub page (and likely the one reviewed), and the difference between 1.0.1 and 1.0.0 is minute.
