### 2018-05-07 #515 @BenjaK
@leeper The DOI for version 1.0.1 is 10.5281/zenodo.1242982. Points 1 and 2 are thus taken care of.
