### 2018-05-09 #515 @leeper
@BenjaK Great. Thank you. The archived version is 1.0.1 and that is the version we will consider to have been reviewed by JOSS. Happy to say that means your paper is now accepted for JOSS! Apologies for the longer than anticipated review timeline.

@arfon Over to you.
