### 2018-05-09 #515 @BenjaK
@leeper Thank you. However, it seems that the article proof is the old version of the article; not the updated version found here: https://github.com/BenjaK/scanstatistics/blob/master/JOSS/paper.md
